third_party/WebKit/LayoutTests/fast/events/init-message-event.html

<!DOCTYPE html>
<html>
<head>
<link rel="help" href="http://www.w3.org/TR/2013/WD-html51-20130528/editing.html#dom-datatransferitemlist-remove">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Checks that initMessageEvent() is working");

var newEvent;
function runTest() {
    newEvent = document.createEvent("MessageEvent");
    shouldThrow('newEvent.initMessageEvent("message", true, true, "ff", "*", 43, window, {x:1});', '"TypeError: The 8th argument is neither an array, nor does it have indexed properties."');
    shouldNotThrow('newEvent.initMessageEvent("message", true, true, "ff", "*", 43, window, null)');
    shouldBeType(newEvent, MessageEvent);
}

runTest();
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
